Add client sub-module
