+++ mithril-relay/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod client;
 pub mod peer;
 pub mod relay;
